ver2

+++ src/main/java/com/skel/repository/AppRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by hootting on 2016. 10. 11..
  */
-public interface AppRepository extends JpaRepository<App,Long> {
+public interface AppRepository extends JpaRepository<App,Integer> {
 }
