+++ src/main/java/com/skel/repository/ChatRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by hootting on 2016. 10. 11..
  */
-public interface ChatRepository extends JpaRepository<Chat,Long> {
+public interface ChatRepository extends JpaRepository<Chat,Integer> {
 }
